Task done true

+++ app/routes.py
@@ -5,6 +5,9 @@ from app.forms import LoginForm, RegistrationForm, BoardForm, CardForm, TaskForm
 from flask_login import current_user, login_user
 from app.models import User, Board, Card, Task
 from flask_login import logout_user, login_required
+from datetime import date
+
+
 
 @app.route("/")
 def index():
@@ -100,65 +103,35 @@ def board(boardid):
     print (thisboard.title)
     print (thisboard.id)
     thisboardid = thisboard.id
-    form = CardForm()
-
-    # form for cards
-    if form.validate_on_submit():
-        cardtaken = Card.query.filter_by(date=form.date.data).filter_by(board_id=thisboardid).first()
-        if cardtaken is None:
-            card = Card(header=form.header.data, date=form.date.data, motherboard=thisboard)
-            db.session.add(card)
-            db.session.commit()
-        # flash('You have only 1 card per day on your Board!')
-        print('Card taken, choose another date')
-        # print(card.id)
-        # print("cardtaken is:")
-        # print(cardtaken)
-   
-
-    # form for tasks
-    formTask = TaskForm()
-    if formTask.validate_on_submit():
-        task = Task(card_id=formTask.card_id.data, tasktext=formTask.tasktext.data)
-        db.session.add(task)
-        db.session.commit()
-        return redirect(url_for('board', boardid=boardid))
-
     # display all card on this board________________
     cards = thisboard.cards.order_by(Card.date).all()
+    print (cards)
     #display all tasks______________________________
     tasks = Task.query.all()
-
-    # for card in cards:
-    #     # if card.id > 55:
-    #     #     print (card.id)
-    #     print (card.header)
-    #     cardid= card.id
-    #     print("tohle je card ID")
-    #     print(cardid)
-    # allcardsonboard = Card.query.filter_by(board_id=thisboardid).all()
-    # print(allcardsonboard)
-    # count = Card.query.filter_by(board_id=thisboardid).count()
-    # print(count)
-
-     # Delete task form
+    # forms__________________________________________
+    form = CardForm()
+    formTask = TaskForm()
     formDeleteTask = DeleteTaskForm()
-    # Done task form
     formDoneTask = DoneTaskForm()
-    # Done task form
-    # formUnDoneTask = UnDoneTaskForm()
-     # Delete card form
     formDeleteCard = DeleteCardForm()
    
-
-    
-    # print (tasks)
-    print (cards)
-
-    # Delete task form
+    # forms___________________________________________
     if request.method =='POST':
         form_name = request.form['form-name']
-        if form_name == 'form-taskDelete':
+        if form_name == 'form-newCard':
+            if form.submitc and form.validate():
+                cardtaken = Card.query.filter_by(date=form.date.data).filter_by(board_id=thisboardid).first()
+                if cardtaken is None:
+                    card = Card(header=form.header.data, date=form.date.data, motherboard=thisboard)
+                    db.session.add(card)
+                    db.session.commit()
+        elif form_name == 'form-newTask':
+            if formTask.validate_on_submit():
+                task = Task(card_id=formTask.card_id.data, tasktext=formTask.tasktext.data)
+                db.session.add(task)
+                db.session.commit()
+                return redirect(url_for('board', boardid=boardid))
+        elif form_name == 'form-taskDelete':
             if formDeleteTask.validate_on_submit():
                 taskToDelete = Task.query.filter_by(id=formDeleteTask.id.data).first()
                 db.session.delete(taskToDelete)
@@ -178,21 +151,53 @@ def board(boardid):
                     db.session.commit()
                     return redirect(url_for('board', boardid=boardid))
                 print(taskDone)
+        elif form_name == 'form-cardDelete':   
+            # Delete card form
+            if formDeleteCard.validate_on_submit():
+                if Task.query.filter_by(card_id=formDeleteCard.id.data).first() is None:
+                    cardToDelete = Card.query.filter_by(id=formDeleteCard.id.data).first()
+                    db.session.delete(cardToDelete)
+                    db.session.commit()
+                    return redirect(url_for('board', boardid=boardid))
+                for task in Task.query.filter_by(card_id=formDeleteCard.id.data).all():
+                    #problem with deleting tasks
+                    db.session.delete(task)
+                    db.session.commit()
+                    cardToDelete = Card.query.filter_by(id=formDeleteCard.id.data).first()
+                    db.session.delete(cardToDelete)
+                    db.session.commit()
+                return redirect(url_for('board', boardid=boardid))
+    # dashboard____________________________________________________________________________
+    today = str(date.today())
+    allTasksForToday=0
+    doneTasksForToday=0
+    rainbowValue= None
+    rainbowMeter=350
+    import random
+    #CARD
+    if Card.query.filter_by(date=today).filter_by(board_id=thisboardid).first():
+        cardForToday = Card.query.filter_by(date=today).filter_by(board_id=thisboardid).first()
+        #TODAYS TASKS
+        if Task.query.filter_by(card_id=cardForToday.id).first():
+            allTasksForToday = Task.query.filter_by(card_id=cardForToday.id).all()
+            numberAllTasks = len(allTasksForToday)
+            #DONE TODAYS TASKS
+            doneTasksForToday = Task.query.filter_by(card_id=cardForToday.id).filter_by(done=True).all()
+            numberDoneTasks = len(doneTasksForToday)    
+            #RAINBOW VALUE
+            rainbowValue = numberDoneTasks/numberAllTasks
+            rainbowMeter = 350-350*rainbowValue
 
+            print("Todays card is:",cardForToday.header)
+            print ("Number of all tasks = ", numberAllTasks)
+            print ("Number of done tasks = ", numberDoneTasks)
+            print ("Rainbow value = ", rainbowValue)
+            
+            print ("Random choice = ", random.choice(allTasksForToday))
 
-    # Delete card form
-    if formDeleteCard.validate_on_submit():
-        if Task.query.filter_by(card_id=formDeleteCard.id.data).all() is None:
-            cardToDelete = Card.query.filter_by(id=formDeleteCard.id.data).first()
-            db.session.delete(cardToDelete)
-            db.session.commit()
-            return redirect(url_for('board', boardid=boardid))
-        for task in Task.query.filter_by(card_id=formDeleteCard.id.data).all():
-            db.session.delete(task)
-            db.session.commit()
-            db.session.delete(cardToDelete)
-            db.session.commit()
-        return redirect(url_for('board', boardid=boardid))
+            if cardForToday:
+                dashboardAwardTitle = "Get award for today!"
+
 
 
     # for card in cards:
@@ -207,11 +212,24 @@ def board(boardid):
     #         print(tasksOnCard)
 
 
+  # for card in cards:
+    #     # if card.id > 55:
+    #     #     print (card.id)
+    #     print (card.header)
+    #     cardid= card.id
+    #     print("tohle je card ID")
+    #     print(cardid)
+    # allcardsonboard = Card.query.filter_by(board_id=thisboardid).all()
+    # print(allcardsonboard)
+    # count = Card.query.filter_by(board_id=thisboardid).count()
+    # print(count)
     
 
     
 
-    return render_template('thisboard.html', user=user, title=thisboard.title, greeting="Let's do it!", boards=boards, form=form, cards=cards, formTask=formTask, tasks=tasks, formDeleteTask=formDeleteTask, formDeleteCard=formDeleteCard, formDoneTask=formDoneTask)
+    return render_template('thisboard.html', user=user, title=thisboard.title, greeting="Let's do it!", boards=boards, form=form, cards=cards, 
+    formTask=formTask, tasks=tasks, formDeleteTask=formDeleteTask, formDeleteCard=formDeleteCard, formDoneTask=formDoneTask, allTasksForToday=allTasksForToday,
+    doneTasksForToday=doneTasksForToday, rainbowValue=rainbowValue, rainbowMeter=rainbowMeter)
 
 
 @app.route("/test")
